Split AsyncTokenBucket into individual source code files
