+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/qos/AsyncTokenBucketBuilder.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.pulsar.broker.qos;
+
+// CHECKSTYLE.OFF: ClassTypeParameterName
+public abstract class AsyncTokenBucketBuilder<SELF extends AsyncTokenBucketBuilder<SELF>> {
+    protected MonotonicClockSource clockSource = AsyncTokenBucket.DEFAULT_CLOCK_SOURCE;
+    protected long resolutionNanos = AsyncTokenBucket.defaultResolutionNanos;
+
+    protected AsyncTokenBucketBuilder() {
+    }
+
+    protected SELF self() {
+        return (SELF) this;
+    }
+
+    public SELF clockSource(MonotonicClockSource clockSource) {
+        this.clockSource = clockSource;
+        return self();
+    }
+
+    public SELF resolutionNanos(long resolutionNanos) {
+        this.resolutionNanos = resolutionNanos;
+        return self();
+    }
+
+    public abstract AsyncTokenBucket build();
+}
